fix(windows): make MIR run on MinGW

+++ c-tests/new/bf2.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-typedef unsigned long uint64_t;
+typedef unsigned long long uint64_t;
 struct S1 {
   volatile signed f0 : 26;
   volatile uint64_t f1;
@@ -8,8 +8,8 @@ struct S1 {
   volatile signed f3 : 1;
 };
 
-static struct S1 g_90 = {72, 18446744073709551615UL, 46, -0}; /* VOLATILE GLOBAL g_90 */
+static struct S1 g_90 = {72, 18446744073709551615ULL, 46, -0}; /* VOLATILE GLOBAL g_90 */
 int main (void) {
-  printf ("%d,%lu,%d,%d(size=%d)\n", g_90.f0, g_90.f1, g_90.f2, g_90.f3, sizeof (struct S1));
+  printf ("%d,%llu,%d,%d(size=%d)\n", g_90.f0, g_90.f1, g_90.f2, g_90.f3, sizeof (struct S1));
   return 0;
 }
